backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
fastapi==0.128.0
ffmpeg-python==0.2.0
flatbuffers==25.12.19
fonttools==4.61.1
future==1.0.0
google-auth==2.47.0
google-genai==1.59.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
mediapipe==0.10.32
numpy==2.4.2
opencv-contrib-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
starlette==0.50.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==15.0.1
youtube-transcript-api==1.2.4
yt-dlp==2026.2.4
